requirements.txt
asgiref==3.8.1
attrs==24.2.0
dj-database-url==2.2.0
Django==5.0.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
whitenoise==6.7.0
